Introduction of Prepare support for kiwi-ng
