+++ src/berry_mill/kiwiapp.py
@@ -2,15 +2,17 @@ import os
 import sys
 from typing import Dict, List
 from abc import ABC, abstractmethod
-from berry_mill.boxbuild import BoxBuildTask
 
+from berry_mill.boxbuild import BoxBuildTask
 from berry_mill.localwrap import LocalBuildTask
 import pathlib
+from berry_mill.preparetask import PrepareTask
 
 class KiwiApp(ABC):
     """
     Abstract Base Class 
     Posing as a costum API for calling kiwi programmatically
+    Functions as Interface between Berrymill and Kiwi-ng Wrappers
     """
     @abstractmethod
     def __init__(self, argv:List[str], repos:Dict[str, Dict[str,str]]):
@@ -21,8 +23,21 @@ class KiwiApp(ABC):
     def run(self) -> None:
         pass
 
+class KiwiAppPrepare(KiwiApp):
+    """
+    Interface between Berrymill and Kiwi-ng Prepare Wrapper
+    """
+    def __init__(self, argv:List[str], repos:Dict[str, Dict[str,str]]):
+        super().__init__(argv, repos)
+    
+    def run(self) -> None:
+        PrepareTask(self._repos).process()
+
 
 class KiwiAppLocal(KiwiApp):
+    """
+    Interface between Berrymill and Kiwi-ng build Wrapper
+    """
 
     def __init__(self, argv:List[str], repos:Dict[str, Dict[str,str]]):
         super().__init__(argv, repos)
@@ -40,7 +55,9 @@ class KiwiAppLocal(KiwiApp):
 
 
 class KiwiAppBox(KiwiApp):
-    
+    """
+    Interface between Berrymill and Kiwi-ng boxbuild Wrapper
+    """
     def __init__(self, argv:List[str], repos:Dict[str, Dict[str,str]], args_tmp_dir:str = None):
         super().__init__(argv, repos)
